feat(validator): humanized validation messages
- refactor: ErrorHumanizer

closes #22
