+++ packages/validator/src/humanizer/formatters/Formatter.ts
@@ -1,0 +1,5 @@
+import { ErrorObject } from 'ajv';
+
+export interface Formatter {
+  format(error: ErrorObject): string;
+}
