--- packages/validator/src/humanizer/formatters/Helpers.ts
@@ -1,18 +1,0 @@
-/**
- * returns a human-readable version of the given list of values.
- */
-export const humanizeList = (arr: string[], conjunction = 'and'): string => {
-  if (arr.length === 0) {
-    return 'nothing';
-  }
-  if (arr.length === 1) {
-    return arr[0];
-  }
-  if (arr.length === 2) {
-    return `${arr[0]} ${conjunction} ${arr[1]}`;
-  }
-
-  return `${arr.slice(0, -1).join(', ')}, ${conjunction} ${
-    arr[arr.length - 1]
-  }`;
-};
